requirements.txt
pycrypto==2.6.1
pycryptodome==3.9.0